Build fix

+++ swift/Sources/imperium-server/Mutex.swift
@@ -7,7 +7,7 @@ public class Mutex {
     public init() {
         var attr: pthread_mutexattr_t = pthread_mutexattr_t()
         pthread_mutexattr_init(&attr)
-        pthread_mutexattr_settype(&attr, PTHREAD_MUTEX_RECURSIVE)
+        pthread_mutexattr_settype(&attr, Int32(PTHREAD_MUTEX_RECURSIVE))
 
         let err = pthread_mutex_init(&self.mutex, &attr)
         pthread_mutexattr_destroy(&attr)
